I don't think it makes sense for `spider_cache` to be a global group.
Every site in the installation should be able to clear the output cache content specific to that site. If all page caches need to be invalidated, I think it's safe to assume that something bigger happened requiring that drastic of a measure.

I believe this followed Batcache's lead originally.

We'll ship this in today's update, and roll back if there are any unintended consequences.